refresh-device-update-with-latest-rlc-codegen
