+++ sdk/deviceupdate/iot-device-update-rest/api-extractor.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-  "mainEntryPointFilePath": "types/src/index.d.ts",
+  "mainEntryPointFilePath": "./types/src/index.d.ts",
   "docModel": {
     "enabled": true
   },
@@ -11,7 +11,7 @@
   "dtsRollup": {
     "enabled": true,
     "untrimmedFilePath": "",
-    "publicTrimmedFilePath": "./types/iot-device-update-rest.d.ts"
+    "publicTrimmedFilePath": "./types/iot-device-update.d.ts"
   },
   "messages": {
     "tsdocMessageReporting": {
