+++ sdk/deviceupdate/iot-device-update-rest/package.json
@@ -2,8 +2,8 @@
   "name": "@azure-rest/iot-device-update",
   "sdk-type": "client",
   "author": "Microsoft Corporation",
-  "description": "Device Update for IoT Hub is an Azure service that enables customers to publish update for their IoT devices to the cloud, and then deploy that update to their devices (approve updates to groups of devices managed and provisioned in IoT Hub). It leverages the proven security and reliability of the Windows Update platform, optimized for IoT devices. It works globally and knows when and how to update devices, enabling customers to focus on their business goals and let Device Update for IoT Hub handle the updates.",
   "version": "1.0.0-beta.1",
+  "description": "",
   "keywords": [
     "node",
     "azure",
@@ -13,10 +13,10 @@
     "isomorphic"
   ],
   "license": "MIT",
-  "main": "./dist/index.js",
+  "main": "dist/index.js",
   "module": "./dist-esm/src/index.js",
-  "types": "./types/iot-device-update-rest.d.ts",
-  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/deviceupdate/iot-device-update/README.md",
+  "types": "./types/iot-device-update.d.ts",
+  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/deviceupdate/iot-device-update-rest/README.md",
   "repository": "github:Azure/azure-sdk-for-js",
   "bugs": {
     "url": "https://github.com/Azure/azure-sdk-for-js/issues"
@@ -24,13 +24,11 @@
   "files": [
     "dist/",
     "dist-esm/src/",
-    "types/iot-device-update-rest.d.ts",
+    "types/iot-device-update.d.ts",
     "README.md",
-    "LICENSE"
+    "LICENSE",
+    "review/*"
   ],
-  "engines": {
-    "node": ">=12.0.0"
-  },
   "//metadata": {
     "constantPaths": [
       {
@@ -39,27 +37,15 @@
       }
     ]
   },
-  "//sampleConfiguration": {
-    "productName": "Azure IoT Hub Device Update rest",
-    "productSlugs": [
-      "azure"
-    ],
-    "requiredResources": {
-      "Azure Iot Device Update instance": "https://docs.microsoft.com/azure/iot-hub-device-update/understand-device-update"
-    },
-    "disableDocsMs": true,
-    "apiRefLink": "https://docs.microsoft.com/rest/api/deviceupdate/2021-06-01-preview/device-update"
-  },
-  "browser": {
-    "./dist-esm/test/public/utils/env.js": "./dist-esm/test/public/utils/env.browser.js"
+  "engines": {
+    "node": ">=12.0.0"
   },
   "scripts": {
     "audit": "node ../../../common/scripts/rush-audit.js && rimraf node_modules package-lock.json && npm i --package-lock-only 2>&1 && npm audit",
-    "build:browser": "tsc -p . && dev-tool run bundle",
-    "build:node": "tsc -p . && dev-tool run bundle",
-    "build:samples": "echo Obsolete.",
+    "build:browser": "tsc -p . && cross-env ONLY_BROWSER=true rollup -c 2>&1",
+    "build:node": "tsc -p . && cross-env ONLY_NODE=true rollup -c 2>&1",
+    "build:samples": "echo skipped.",
     "build:test": "tsc -p . && dev-tool run bundle",
-    "build": "npm run clean && tsc -p . && dev-tool run bundle && mkdirp ./review && api-extractor run --local",
     "build:debug": "tsc -p . && dev-tool run bundle && api-extractor run --local",
     "check-format": "prettier --list-different --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"samples-dev/**/*.ts\" \"*.{js,json}\"",
     "clean": "rimraf dist dist-browser dist-esm test-dist temp types *.tgz *.log",
@@ -76,56 +62,66 @@
     "test:browser": "npm run clean && npm run build:test && npm run unit-test:browser",
     "test:node": "npm run clean && npm run build:test && npm run unit-test:node",
     "test": "npm run clean && npm run build:test && npm run unit-test",
-    "unit-test:browser": "dev-tool run test:browser",
+    "unit-test": "npm run unit-test:node && npm run unit-test:browser",
     "unit-test:node": "dev-tool run test:node-ts-input -- --timeout 1200000 --exclude 'test/**/browser/*.spec.ts' 'test/**/*.spec.ts'",
-    "unit-test": "npm run unit-test:node && npm run unit-test:browser"
+    "unit-test:browser": "dev-tool run test:browser",
+    "build": "npm run clean && tsc -p . && dev-tool run bundle && mkdirp ./review && api-extractor run --local"
   },
   "sideEffects": false,
   "autoPublish": false,
   "dependencies": {
     "@azure/core-auth": "^1.3.0",
-    "@azure-rest/core-client": "1.0.0-beta.9",
-    "@azure/core-rest-pipeline": "^1.1.0",
+    "@azure-rest/core-client": "1.0.0-beta.10",
+    "@azure/core-rest-pipeline": "^1.8.0",
     "@azure/logger": "^1.0.0",
     "tslib": "^2.2.0",
-    "@azure/core-paging": "^1.1.1",
+    "@azure/core-paging": "^1.2.0",
     "@azure/core-lro": "^2.2.0"
   },
   "devDependencies": {
+    "@microsoft/api-extractor": "7.18.11",
+    "autorest": "latest",
+    "@types/node": "^12.0.0",
+    "dotenv": "^8.2.0",
+    "eslint": "^7.15.0",
+    "mkdirp": "^1.0.4",
+    "prettier": "2.2.1",
+    "rimraf": "^3.0.0",
+    "source-map-support": "^0.5.9",
+    "typescript": "~4.2.0",
     "@azure/dev-tool": "^1.0.0",
     "@azure/eslint-plugin-azure-sdk": "^3.0.0",
+    "@azure-tools/test-credential": "^1.0.0",
     "@azure/identity": "^2.0.1",
     "@azure-tools/test-recorder": "^2.0.0",
-    "@azure-tools/test-credential": "1.0.0",
-    "@microsoft/api-extractor": "7.18.11",
-    "@types/chai": "^4.1.6",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^12.0.0",
-    "chai": "^4.2.0",
+    "mocha": "^7.1.1",
+    "mocha-junit-reporter": "^1.18.0",
     "cross-env": "^7.0.2",
-    "dotenv": "^8.2.0",
-    "eslint": "^8.0.0",
+    "@types/chai": "^4.2.8",
+    "chai": "^4.2.0",
     "karma-chrome-launcher": "^3.0.0",
     "karma-coverage": "^2.0.0",
     "karma-edge-launcher": "^0.4.2",
     "karma-env-preprocessor": "^0.1.1",
     "karma-firefox-launcher": "^1.1.0",
     "karma-ie-launcher": "^1.0.0",
-    "karma-json-preprocessor": "^0.3.3",
-    "karma-json-to-file-reporter": "^1.0.1",
     "karma-junit-reporter": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-mocha": "^2.0.1",
     "karma-source-map-support": "~1.4.0",
     "karma-sourcemap-loader": "^0.3.8",
     "karma": "^6.2.0",
-    "mkdirp": "^1.0.4",
-    "mocha-junit-reporter": "^1.18.0",
-    "mocha": "^7.1.1",
-    "nyc": "^15.0.0",
-    "prettier": "2.2.1",
-    "rimraf": "^3.0.0",
-    "source-map-support": "^0.5.9",
-    "typescript": "~4.2.0"
+    "nyc": "^14.0.0"
+  },
+  "browser": {
+    "./dist-esm/test/public/utils/env.js": "./dist-esm/test/public/utils/env.browser.js"
+  },
+  "//sampleConfiguration": {
+    "productName": "",
+    "productSlugs": [
+      "azure"
+    ],
+    "disableDocsMs": true,
+    "apiRefLink": "https://docs.microsoft.com/javascript/api/@azure-rest/iot-device-update?view=azure-node-preview"
   }
 }
