+++ sdk/deviceupdate/iot-device-update-rest/sample.env
@@ -1,7 +1,4 @@
-# Iot Device Update resource endpoint
-ENDPOINT=
-INSTANCE_ID=
 # App registration secret for AAD authentication
 AZURE_CLIENT_SECRET=
 AZURE_CLIENT_ID=
-AZURE_TENANT_ID=
+AZURE_TENANT_ID=
