+++ sdk/deviceupdate/iot-device-update-rest/src/index.ts
@@ -6,6 +6,8 @@ import DeviceUpdate from "./deviceUpdate";
 export * from "./deviceUpdate";
 export * from "./parameters";
 export * from "./responses";
+export * from "./clientDefinitions";
+export * from "./isUnexpected";
 export * from "./models";
 export * from "./outputModels";
 export * from "./paginateHelper";
