+++ sdk/deviceupdate/iot-device-update-rest/src/models.ts
@@ -59,7 +59,7 @@ export interface Deployment {
   /** The deployment identifier. */
   deploymentId: string;
   /** The deployment start datetime. */
-  startDateTime: string;
+  startDateTime: Date | string;
   /** Update identity. */
   updateId: UpdateId;
   /** The group identity */
