+++ sdk/deviceupdate/iot-device-update-rest/src/paginateHelper.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { PagedAsyncIterableIterator, PagedResult, getPagedAsyncIterator } from "@azure/core-paging";
-import { Client, PathUncheckedResponse, createRestError } from "@azure-rest/core-client";
+import { getPagedAsyncIterator, PagedAsyncIterableIterator, PagedResult } from "@azure/core-paging";
+import { Client, createRestError, PathUncheckedResponse } from "@azure-rest/core-client";
 
 /**
  * Helper type to extract the type of an array
