+++ sdk/deviceupdate/iot-device-update-rest/src/pollingHelper.ts
@@ -7,8 +7,8 @@ import {
   LroEngine,
   LroEngineOptions,
   LroResponse,
-  PollOperationState,
   PollerLike,
+  PollOperationState,
 } from "@azure/core-lro";
 
 /**
@@ -38,7 +38,9 @@ export function getLongRunningPoller<TResult extends HttpResponse>(
       // which is an opaque URL provided by caller, the service sends this in one of the following headers: operation-location, azure-asyncoperation or location
       // depending on the lro pattern that the service implements. If non is provided we default to the initial path.
       const response = await client.pathUnchecked(path ?? initialResponse.request.url).get();
-      return getLroResponse(response as TResult);
+      const lroResponse = getLroResponse(response as TResult);
+      lroResponse.rawResponse.headers["x-ms-original-url"] = initialResponse.request.url;
+      return lroResponse;
     },
   };
