+++ sdk/deviceupdate/iot-device-update-rest/test/public/utils/env.browser.ts
@@ -1,0 +1,2 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
