+++ sdk/deviceupdate/iot-device-update-rest/test/public/utils/recordedClient.ts
@@ -1,11 +1,37 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { createTestCredential } from "@azure-tools/test-credential";
-import DeviceUpdate, { DeviceUpdateRestClient } from "../../../src";
-import { Recorder, assertEnvironmentVariable } from "@azure-tools/test-recorder";
+import { Context } from "mocha";
+import { assertEnvironmentVariable, Recorder, RecorderStartOptions } from "@azure-tools/test-recorder";
+import { createTestCredential } from '@azure-tools/test-credential'
+import DeviceUpdate, { DeviceUpdateClient } from "../../../src/index";
+import "./env";
 
-export function createRecordedClient(recorder: Recorder): DeviceUpdateRestClient {
+const envSetupForPlayback: Record<string, string> = {
+  ENDPOINT: "https://endpoint",
+  AZURE_CLIENT_ID: "azure_client_id",
+  AZURE_CLIENT_SECRET: "azure_client_secret",
+  AZURE_TENANT_ID: "88888888-8888-8888-8888-888888888888",
+  SUBSCRIPTION_ID: "azure_subscription_id",
+};
+
+const recorderEnvSetup: RecorderStartOptions = {
+  envSetupForPlayback,
+};
+
+/**
+ * creates the recorder and reads the environment variables from the `.env` file.
+ * Should be called first in the test suite to make sure environment variables are
+ * read before they are being used.
+ */
+export async function createRecorder(context: Context): Promise<Recorder> {
+  const recorder = new Recorder(context.currentTest);
+  await recorder.start(recorderEnvSetup);
+  return recorder;
+}
+
+
+export function createRecordedClient(recorder: Recorder): DeviceUpdateClient {
   const credential = createTestCredential();
   const client = DeviceUpdate(
     assertEnvironmentVariable("ENDPOINT"),
@@ -13,15 +39,4 @@ export function createRecordedClient(recorder: Recorder): DeviceUpdateRestClient
     recorder.configureClientOptions({})
   );
   return client;
-}
-
-export async function startRecorder(recorder: Recorder): Promise<void> {
-  await recorder.start({
-    envSetupForPlayback: {
-      ENDPOINT: "endpoint",
-      AZURE_CLIENT_ID: "azure_client_id",
-      AZURE_CLIENT_SECRET: "azure_client_secret",
-      AZURE_TENANT_ID: "88888888-8888-8888-8888-888888888888",
-    },
-  });
 }
